requirements.txt
Django~=3.1
boto3~=1.15.18
requests~=2.21.0
api~=0.0.7
djangorestframework~=3.12.1